Setup realtime/cache servers and utilities
This is a huge commit that serves more as a savepoint. At this stage, both the API server and the Realtime server are running and communicating. A new MySQL Provider base class is added in utilities, to be used by all servers. The Provider for the API server is renamed and derives said base class. The Options class is now a common utility to be used by all servers. Added processing of flags for the Realtime server. Created config.ini files for realtime/cache servers. Added verification of request headers. Initial concept of MySQL flights table.